Add initial database schema migration
Introduce migration `20250404190301_CriandoBanco` to create the initial database schema. This includes adding `Categorias`, `Statuses`, and `Tarefas` tables with appropriate columns and foreign key constraints. Initial data is seeded for `Categorias` and `Statuses`. Indexes are created on `CategoriaId` and `StatusId` in `Tarefas`. The `Down` method is implemented to drop the tables if rolled back. Auto-generated files are updated to reflect these changes.